+++ packages/shared/assets/IoEyeOff.tsx
@@ -1,0 +1,23 @@
+import { ComponentProps } from 'react';
+
+export default function IoEyeOff({ ...props }: ComponentProps<'svg'>) {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M13.5002 14.0007C13.4346 14.0008 13.3695 13.9879 13.3089 13.9628C13.2482 13.9376 13.1931 13.9007 13.1468 13.8542L2.14679 2.85415C2.05698 2.75962 2.00765 2.63375 2.00932 2.50336C2.01099 2.37298 2.06352 2.24841 2.15572 2.15621C2.24792 2.06401 2.3725 2.01148 2.50288 2.00981C2.63326 2.00814 2.75913 2.05747 2.85366 2.14728L13.8537 13.1473C13.9235 13.2172 13.9711 13.3063 13.9904 13.4032C14.0097 13.5002 13.9998 13.6007 13.9619 13.692C13.9241 13.7834 13.8601 13.8614 13.7779 13.9164C13.6957 13.9713 13.5991 14.0007 13.5002 14.0007V14.0007ZM7.75023 9.87103L6.13179 8.25259C6.12251 8.24339 6.11059 8.23733 6.0977 8.23526C6.0848 8.23319 6.07158 8.23521 6.05989 8.24103C6.04821 8.24686 6.03864 8.2562 6.03253 8.26775C6.02642 8.27929 6.02409 8.29246 6.02585 8.3054C6.09119 8.72529 6.28834 9.11352 6.58882 9.414C6.8893 9.71447 7.27753 9.91163 7.69741 9.97696C7.71035 9.97873 7.72352 9.97639 7.73507 9.97028C7.74661 9.96418 7.75595 9.95461 7.76178 9.94292C7.76761 9.93123 7.76963 9.91801 7.76756 9.90512C7.76548 9.89222 7.75942 9.8803 7.75023 9.87103ZM8.25023 6.1304L9.87116 7.75071C9.88041 7.76004 9.89237 7.76621 9.90533 7.76836C9.91829 7.77051 9.93159 7.76851 9.94335 7.76267C9.95512 7.75682 9.96474 7.74742 9.97085 7.73579C9.97697 7.72416 9.97926 7.71091 9.97741 7.6979C9.91224 7.27745 9.71494 6.88865 9.41409 6.58779C9.11323 6.28693 8.72443 6.08964 8.30398 6.02446C8.29096 6.02246 8.27763 6.02461 8.26591 6.03063C8.25419 6.03664 8.24467 6.04621 8.23871 6.05795C8.23275 6.0697 8.23065 6.08303 8.23272 6.09605C8.23479 6.10906 8.24092 6.12108 8.25023 6.1304V6.1304Z"
+        fill="black"
+      />
+      <path
+        d="M15.3437 8.5425C15.4472 8.38012 15.5019 8.19146 15.5014 7.99892C15.5008 7.80638 15.445 7.61804 15.3406 7.45625C14.5137 6.1775 13.4409 5.09125 12.2384 4.31469C10.9062 3.45469 9.4375 3 7.99 3C7.22691 3.00105 6.469 3.12547 5.74562 3.36844C5.72538 3.37517 5.70719 3.38698 5.6928 3.40273C5.6784 3.41848 5.66828 3.43765 5.6634 3.45842C5.65851 3.47919 5.65902 3.50086 5.66488 3.52138C5.67074 3.5419 5.68175 3.56057 5.69687 3.57562L7.17312 5.05188C7.18846 5.06724 7.20753 5.07834 7.22846 5.0841C7.24939 5.08985 7.27146 5.09005 7.2925 5.08469C7.79289 4.96274 8.31625 4.97168 8.81219 5.11063C9.30812 5.24959 9.75993 5.51388 10.1241 5.87807C10.4883 6.24225 10.7526 6.69406 10.8916 7.19C11.0305 7.68594 11.0394 8.2093 10.9175 8.70969C10.9122 8.73068 10.9124 8.75268 10.9182 8.77355C10.9239 8.79443 10.935 8.81344 10.9503 8.82875L13.0737 10.9537C13.0958 10.9759 13.1254 10.9889 13.1567 10.9903C13.1879 10.9917 13.2185 10.9813 13.2425 10.9613C14.0599 10.2646 14.7682 9.44924 15.3437 8.5425ZM8 11C7.54586 11 7.09764 10.8969 6.68913 10.6985C6.28063 10.5001 5.9225 10.2115 5.64175 9.85458C5.361 9.49761 5.16496 9.08156 5.06841 8.6378C4.97186 8.19404 4.97732 7.73416 5.08437 7.29281C5.08968 7.27182 5.08945 7.24982 5.0837 7.22895C5.07795 7.20807 5.06687 7.18906 5.05156 7.17375L2.96281 5.08406C2.94066 5.06189 2.91099 5.04884 2.87968 5.04751C2.84837 5.04618 2.8177 5.05666 2.79375 5.07688C2.03156 5.72719 1.325 6.51844 0.679061 7.44312C0.565994 7.60541 0.503723 7.79764 0.500162 7.9954C0.496601 8.19315 0.55191 8.3875 0.659061 8.55375C1.48437 9.84531 2.54625 10.9331 3.73031 11.6991C5.06437 12.5625 6.4975 13 7.99 13C8.76079 12.9979 9.52658 12.8762 10.26 12.6391C10.2804 12.6325 10.2988 12.6209 10.3134 12.6052C10.3281 12.5895 10.3384 12.5703 10.3434 12.5494C10.3485 12.5286 10.3481 12.5068 10.3423 12.4862C10.3365 12.4655 10.3255 12.4467 10.3103 12.4316L8.82687 10.9484C8.81157 10.9331 8.79255 10.9221 8.77168 10.9163C8.75081 10.9105 8.7288 10.9103 8.70781 10.9156C8.47606 10.9718 8.23845 11.0001 8 11V11Z"
+        fill="black"
+      />
+    </svg>
+  );
+}
